refactor: controller 로직 Facade로 이전
